refactor: remove deprecated ClientPage components and update imports
