+++ src/app/(main)/about/team/page.tsx
@@ -1,9 +1,8 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import TeamPageContent from "@/components/about/TeamPageContent";
 
 export const metadata: Metadata = buildPageMetadata("/about/team");
-
 export default function TeamPage() {
-  return <ClientPage />;
+  return <TeamPageContent />;
 }
