+++ src/app/(main)/contact/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/contact");
 
 export default function ContactPage() {
-  return <ClientPage />;
+  return <Content />;
 }
