+++ src/app/(main)/donate/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import DonatePageContent from "@/components/donate/DonatePageContent";
 
 export const metadata: Metadata = buildPageMetadata("/donate");
 
 export default function DonatePage() {
-  return <ClientPage />;
+  return <DonatePageContent />;
 }
