+++ src/app/(main)/news/Content.tsx
@@ -1,37 +1,15 @@
-"use client";
-
-import { useState } from "react";
 import Image from "@/components/common/Image";
 import Link from "next/link";
 import { ArrowRight, Megaphone } from "lucide-react";
 import Breadcrumb from "@/components/layout/Breadcrumb";
 import { Button } from "@/components/ui/button";
-import { motion } from "framer-motion";
-import { motionDurations } from "@/lib/motion/tokens";
-import {
-  FadeIn,
-  FadeInStagger,
-  FadeInStaggerItem,
-} from "@/components/animations";
-import { AnnouncementCard } from "@/components/common/AnnouncementCard";
-import { getNewsOnlyItems, newsItemToAnnouncement } from "@/data/unified-news";
-import { cn } from "@/lib/utils";
+import { MotionDiv, MotionH1, MotionP } from "@/components/motion/framer";
+import NewsCategoryListing from "@/components/news/NewsCategoryListing";
+import { getNewsOnlyItems } from "@/data/unified-news";
 
 const allItems = getNewsOnlyItems();
 
-const categories = [
-  "All",
-  ...Array.from(new Set(allItems.map((item) => item.category))),
-];
-
-export default function NewsPage() {
-  const [activeCategory, setActiveCategory] = useState("All");
-
-  const filtered =
-    activeCategory === "All"
-      ? allItems
-      : allItems.filter((item) => item.category === activeCategory);
-
+export default function Content() {
   return (
     <>
       <div className="site-container mx-auto">
@@ -46,7 +24,7 @@ export default function NewsPage() {
 
           <div className="site-container mx-auto px-4 sm:px-6 relative z-10 h-full flex flex-col lg:flex-row lg:items-center gap-10 sm:gap-12">
             <div className="flex-1 max-w-3xl pt-4 sm:pt-8 lg:pt-0">
-              <motion.div
+              <MotionDiv
                 initial={{ opacity: 0, y: 20 }}
                 animate={{ opacity: 1, y: 0 }}
                 transition={{ duration: 0.5 }}
@@ -56,9 +34,9 @@ export default function NewsPage() {
                 <span className="text-sm font-medium text-[#eeba2b]">
                   Latest news
                 </span>
-              </motion.div>
+              </MotionDiv>
 
-              <motion.h1
+              <MotionH1
                 initial={{ opacity: 0, y: 40 }}
                 animate={{ opacity: 1, y: 0 }}
                 transition={{ duration: 0.8, ease: "easeOut" }}
@@ -66,9 +44,9 @@ export default function NewsPage() {
               >
                 News from the{" "}
                 <span className="text-[#eeba2b] font-medium">frontlines</span>
-              </motion.h1>
+              </MotionH1>
 
-              <motion.p
+              <MotionP
                 initial={{ opacity: 0, y: 40 }}
                 animate={{ opacity: 1, y: 0 }}
                 transition={{ delay: 0.15, duration: 0.8, ease: "easeOut" }}
@@ -77,9 +55,9 @@ export default function NewsPage() {
                 Awards, partnerships, program launches, and milestones — stay
                 connected with everything shaping the future of student-powered
                 healthcare.
-              </motion.p>
+              </MotionP>
 
-              <motion.div
+              <MotionDiv
                 initial={{ opacity: 0, y: 20 }}
                 animate={{ opacity: 1, y: 0 }}
                 transition={{ delay: 0.3, duration: 0.6 }}
@@ -104,11 +82,11 @@ export default function NewsPage() {
                     <ArrowRight className="ml-2 h-4 w-4 group-hover:translate-x-1 transition-transform" />
                   </Link>
                 </Button>
-              </motion.div>
+              </MotionDiv>
             </div>
 
             {/* Decorative Image */}
-            <motion.div
+            <MotionDiv
               initial={{ opacity: 0, scale: 0.95 }}
               animate={{ opacity: 1, scale: 1 }}
               transition={{ delay: 0.2, duration: 0.8, ease: "easeOut" }}
@@ -144,68 +122,11 @@ export default function NewsPage() {
                   </div>
                 </div>
               </div>
-            </motion.div>
+            </MotionDiv>
           </div>
         </section>
 
-        {/* News grid */}
-        <section className="py-16 md:py-24 bg-[#FCFAEF] dark:bg-[#1C1F1E]">
-          <div className="site-container mx-auto px-4 sm:px-6">
-            <FadeIn
-              className="text-center max-w-3xl mx-auto mb-10"
-              duration={motionDurations.enter}
-            >
-              <h2 className="text-[#F5C94D] font-bold text-base sm:text-lg mb-2">
-                ALL NEWS
-              </h2>
-              <h2 className="text-2xl sm:text-3xl md:text-4xl font-bold mb-4 text-[#1C1F1E] dark:text-[#FCFAEF]">
-                Everything You Need to Know
-              </h2>
-              <p className="text-base sm:text-lg text-[#2F3332]/70 dark:text-[#E6E7E7]/60 leading-relaxed max-w-2xl mx-auto">
-                Browse stories covering awards, partnerships, launches, and
-                milestones from our teams in the field.
-              </p>
-            </FadeIn>
-
-            {/* Category Filter Pills */}
-            <FadeIn className="mb-10" duration={motionDurations.enter}>
-              <div className="flex flex-wrap justify-center gap-2 sm:gap-3 max-w-4xl mx-auto">
-                {categories.map((cat) => (
-                  <button
-                    key={cat}
-                    onClick={() => setActiveCategory(cat)}
-                    className={cn(
-                      "rounded-full px-4 py-2 text-sm font-medium transition-all duration-200",
-                      activeCategory === cat
-                        ? "bg-[#0097b2] text-[#FCFAEF] shadow-md"
-                        : "bg-white dark:bg-[#2F3332] text-[#2F3332]/70 dark:text-[#E6E7E7]/70 border border-black/[0.06] dark:border-white/[0.08] hover:border-[#0097b2]/30 hover:text-[#0097b2] dark:hover:text-[#66C4DC]"
-                    )}
-                  >
-                    {cat}
-                  </button>
-                ))}
-              </div>
-            </FadeIn>
-
-            <FadeInStagger
-              key={activeCategory}
-              className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 sm:gap-8 max-w-6xl mx-auto"
-              staggerDelay={motionDurations.staggerContainer}
-            >
-              {filtered.map((item) => (
-                <FadeInStaggerItem key={item.id} direction="up">
-                  <AnnouncementCard item={newsItemToAnnouncement(item)} />
-                </FadeInStaggerItem>
-              ))}
-            </FadeInStagger>
-
-            {filtered.length === 0 && (
-              <p className="text-center text-[#2F3332]/50 dark:text-[#E6E7E7]/40 mt-12">
-                No stories in this category yet.
-              </p>
-            )}
-          </div>
-        </section>
+        <NewsCategoryListing />
 
         {/* CTA Section */}
         <section className="py-16 md:py-24 bg-gradient-to-r from-[#0097b2] via-[#0F4C5C] to-[#031C3A] text-[#FCFAEF] relative overflow-hidden">
@@ -216,7 +137,7 @@ export default function NewsPage() {
 
           <div className="relative site-container mx-auto px-4 sm:px-6">
             <div className="max-w-3xl mx-auto text-center">
-              <motion.div
+              <MotionDiv
                 initial={{ opacity: 0, y: 24 }}
                 whileInView={{ opacity: 1, y: 0 }}
                 transition={{ duration: 0.6 }}
@@ -252,7 +173,7 @@ export default function NewsPage() {
                     </Link>
                   </Button>
                 </div>
-              </motion.div>
+              </MotionDiv>
             </div>
           </div>
         </section>
