+++ src/app/(main)/privacy/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/privacy");
 
 export default function PrivacyPage() {
-  return <ClientPage />;
+  return <Content />;
 }
