+++ src/app/(main)/programs/akomapa-foods/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/programs/akomapa-foods");
 
-export default function AkomapaFoodsPage() {
-  return <ClientPage />;
+export default function FoodsPage() {
+  return <Content />;
 }
