+++ src/app/(main)/programs/akomapa-ghip/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/programs/akomapa-ghip");
 
-export default function AkomapaGhipPage() {
-  return <ClientPage />;
+export default function GHIPPage() {
+  return <Content />;
 }
