+++ src/app/(main)/programs/akomapa-ghltp/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/programs/akomapa-ghltp");
 
 export default function AkomapaGhltpPage() {
-  return <ClientPage />;
+  return <Content />;
 }
