+++ src/app/(main)/programs/akomapa-network/Content.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { motion } from "framer-motion";
+import { MotionDiv, MotionH1, MotionP } from "@/components/motion/framer";
 import { AnimatedMetric } from "@/components/motion/AnimatedMetric";
 import Link from "next/link";
 import { ArrowLeft, ArrowRight, Quote } from "lucide-react";
@@ -133,7 +131,7 @@ const primaryCtaClass =
 const secondaryCtaClass =
   `${ctaBaseClass} bg-[#eeba2b] hover:bg-[#eeba2b]/80 text-[#FCFAEF] shadow-lg hover:shadow-xl focus-visible:ring-[#F5C94D]`;
 
-export default function AkomapaNetworkPage() {
+export default function Content() {
   return (
     <>
       <div className="site-container mx-auto">
@@ -146,7 +144,7 @@ export default function AkomapaNetworkPage() {
         <div className="absolute bottom-0 left-0 w-96 h-96 bg-[#FCFAEF]/10 rounded-full translate-y-1/2 -translate-x-1/2 blur-3xl" />
 
         <div className="site-container mx-auto px-4 sm:px-6 relative z-10 flex flex-col gap-4 sm:gap-8">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 20 }}
             animate={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -159,33 +157,33 @@ export default function AkomapaNetworkPage() {
               <ArrowLeft className="mr-2 h-4 w-4" />
               Back to Programs
             </Link>
-          </motion.div>
+          </MotionDiv>
           <div className="max-w-5xl">
-            <motion.p
+            <MotionP
               initial={{ opacity: 0, y: 40 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.8, ease: "easeOut" }}
               className="uppercase tracking-[0.3em] text-sm font-semibold text-[#FCFAEF]/80 mb-6"
             >
               The Akomapa Network
-            </motion.p>
-            <motion.h1 
+            </MotionP>
+            <MotionH1 
               initial={{ opacity: 0, y: 40 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.1, duration: 0.8, ease: "easeOut" }}
               className="text-3xl sm:text-4xl md:text-6xl lg:text-7xl font-light text-[#FCFAEF] mb-6 leading-tight"
             >
               Connecting Clinics. Sharing Knowledge. Building the Future of Global Health.
-            </motion.h1>
-            <motion.p 
+            </MotionH1>
+            <MotionP 
               initial={{ opacity: 0, y: 40 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.2, duration: 0.8, ease: "easeOut" }}
               className="text-base sm:text-lg md:text-2xl text-[#FCFAEF]/85 font-light max-w-3xl"
             >
               A global community of student-powered clinics, universities, and mentors working together to reimagine how healthcare is delivered and taught.
-            </motion.p>
-            <motion.div
+            </MotionP>
+            <MotionDiv
             initial={{ opacity: 0, y: 40 }}
             animate={{ opacity: 1, y: 0 }}
             transition={{ delay: 0.3, duration: 0.8, ease: "easeOut" }}
@@ -197,10 +195,10 @@ export default function AkomapaNetworkPage() {
               <Button asChild className={secondaryCtaClass}>
                 <Link href="/partnerships">Partner with Us</Link>
               </Button>
-            </motion.div>
+            </MotionDiv>
           </div>
 
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 40 }}
             animate={{ opacity: 1, y: 0 }}
             transition={{ delay: 0.4, duration: 0.8, ease: "easeOut" }}
@@ -216,7 +214,7 @@ export default function AkomapaNetworkPage() {
               />
               <div className="absolute inset-0 bg-gradient-to-t from-black/35 via-transparent to-transparent" />
             </div>
-          </motion.div>
+          </MotionDiv>
         </div>
       </section>
 
@@ -224,7 +222,7 @@ export default function AkomapaNetworkPage() {
       <section className="py-16 md:py-24 bg-[#FCFAEF] dark:bg-[#1C1F1E]">
         <div className="site-container mx-auto px-4">
           <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 lg:gap-16 items-center">
-            <motion.div
+            <MotionDiv
               initial={{ opacity: 0, y: 24 }}
               whileInView={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.6 }}
@@ -245,8 +243,8 @@ export default function AkomapaNetworkPage() {
                   Our goal is simple but bold: to make community-based, preventative care accessible everywhere while training the next generation of global health leaders.
                 </p>
               </div>
-            </motion.div>
-            <motion.div
+            </MotionDiv>
+            <MotionDiv
               initial={{ opacity: 0, y: 24 }}
               whileInView={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.1, duration: 0.6 }}
@@ -263,7 +261,7 @@ export default function AkomapaNetworkPage() {
                 />
                 <div className="absolute inset-0 bg-gradient-to-t from-black/25 via-transparent to-transparent" />
               </div>
-            </motion.div>
+            </MotionDiv>
           </div>
         </div>
       </section>
@@ -276,7 +274,7 @@ export default function AkomapaNetworkPage() {
         </div>
         <div className="relative site-container mx-auto px-4">
           <div className="max-w-4xl mx-auto">
-            <motion.div
+            <MotionDiv
               initial={{ opacity: 0, y: 24 }}
               whileInView={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.6 }}
@@ -292,7 +290,7 @@ export default function AkomapaNetworkPage() {
               <p className="text-base md:text-lg text-[#FCFAEF]/85 leading-relaxed">
                 We envision a world where a student team in Ghana can learn from one in Chicago, where a New Haven community project can inspire a rural screening model, and where young leaders everywhere are equipped to bridge the gap between knowledge and action.
               </p>
-            </motion.div>
+            </MotionDiv>
           </div>
         </div>
       </section>
@@ -300,7 +298,7 @@ export default function AkomapaNetworkPage() {
       {/* What the Network Does Section */}
       <section className="py-16 md:py-24 bg-[#FCFAEF] dark:bg-[#1C1F1E]">
         <div className="site-container mx-auto px-4">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 24 }}
             whileInView={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -313,11 +311,11 @@ export default function AkomapaNetworkPage() {
             <p className="text-lg text-[#2F3332] dark:text-[#E6E7E7] leading-relaxed">
               Six core functions that connect, empower, and amplify student-powered care across the globe.
             </p>
-          </motion.div>
+          </MotionDiv>
 
           <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 md:gap-8">
             {networkFeatures.map((feature, index) => (
-              <motion.div
+              <MotionDiv
                 key={feature.id}
                 initial={{ opacity: 0, y: 24 }}
                 whileInView={{ opacity: 1, y: 0 }}
@@ -338,7 +336,7 @@ export default function AkomapaNetworkPage() {
                     {feature.description}
                   </p>
                 </div>
-              </motion.div>
+              </MotionDiv>
             ))}
           </div>
         </div>
@@ -351,7 +349,7 @@ export default function AkomapaNetworkPage() {
           <div className="absolute bottom-0 right-0 h-72 w-72 rounded-full bg-[#F5C94D]/10 blur-3xl" />
         </div>
         <div className="relative site-container mx-auto px-4">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 24 }}
             whileInView={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -364,11 +362,11 @@ export default function AkomapaNetworkPage() {
             <p className="text-lg text-[#FCFAEF]/85 leading-relaxed">
               Bold goals that drive our collective mission forward.
             </p>
-          </motion.div>
+          </MotionDiv>
 
           <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 md:gap-8 max-w-6xl mx-auto">
             {goals.map((goal, index) => (
-              <motion.div
+              <MotionDiv
                 key={goal.id}
                 initial={{ opacity: 0, y: 24 }}
                 whileInView={{ opacity: 1, y: 0 }}
@@ -401,7 +399,7 @@ export default function AkomapaNetworkPage() {
                 <p className="text-sm md:text-base text-[#2F3332]/85 leading-relaxed flex-1">
                   {goal.description}
                 </p>
-              </motion.div>
+              </MotionDiv>
             ))}
           </div>
         </div>
@@ -410,7 +408,7 @@ export default function AkomapaNetworkPage() {
       {/* Partner Clinics Section */}
       <section className="py-16 md:py-24 bg-[#FCFAEF] dark:bg-[#1C1F1E]">
         <div className="site-container mx-auto px-4">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 24 }}
             whileInView={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -423,11 +421,11 @@ export default function AkomapaNetworkPage() {
             <p className="text-lg text-[#2F3332] dark:text-[#E6E7E7] leading-relaxed">
               Student-powered clinics leading the way in community-based care and global collaboration.
             </p>
-          </motion.div>
+          </MotionDiv>
 
           <div className="grid grid-cols-1 md:grid-cols-3 gap-8 max-w-6xl mx-auto">
             {partnerClinics.map((clinic, index) => (
-              <motion.div
+              <MotionDiv
                 key={clinic.name}
                 initial={{ opacity: 0, y: 24 }}
                 whileInView={{ opacity: 1, y: 0 }}
@@ -482,7 +480,7 @@ export default function AkomapaNetworkPage() {
                   className="absolute bottom-0 left-0 right-0 h-1 opacity-0 group-hover:opacity-100 transition-opacity"
                   style={{ backgroundColor: clinic.accentColor }}
                 />
-              </motion.div>
+              </MotionDiv>
             ))}
           </div>
         </div>
@@ -496,7 +494,7 @@ export default function AkomapaNetworkPage() {
         </div>
         <div className="relative site-container mx-auto px-4">
           <div className="max-w-4xl mx-auto">
-            <motion.div
+            <MotionDiv
               initial={{ opacity: 0, y: 24 }}
               whileInView={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.6 }}
@@ -506,8 +504,8 @@ export default function AkomapaNetworkPage() {
               <h2 className="text-3xl md:text-4xl font-bold mb-6 leading-tight">
                 Why It Matters
               </h2>
-            </motion.div>
-            <motion.div
+            </MotionDiv>
+            <MotionDiv
               initial={{ opacity: 0, y: 24 }}
               whileInView={{ opacity: 1, y: 0 }}
               transition={{ duration: 0.6, delay: 0.1 }}
@@ -523,7 +521,7 @@ export default function AkomapaNetworkPage() {
               <p className="font-semibold text-[#FCFAEF]">
                 By linking student-powered clinics into one connected system, we&apos;re making community care smarter, stronger, and sustainable — driven by evidence, compassion, and shared purpose.
               </p>
-            </motion.div>
+            </MotionDiv>
           </div>
         </div>
       </section>
@@ -531,7 +529,7 @@ export default function AkomapaNetworkPage() {
       {/* Join the Movement CTA */}
       <section className="py-16 md:py-24 bg-[#FCFAEF] dark:bg-[#1C1F1E]">
         <div className="site-container mx-auto px-4">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 24 }}
             whileInView={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -575,7 +573,7 @@ export default function AkomapaNetworkPage() {
                 </Link>
               </Button>
             </div>
-          </motion.div>
+          </MotionDiv>
         </div>
       </section>
 
@@ -586,7 +584,7 @@ export default function AkomapaNetworkPage() {
           <div className="absolute bottom-0 right-0 h-72 w-72 rounded-full bg-[#F5C94D]/10 blur-3xl" />
         </div>
         <div className="relative site-container mx-auto px-4">
-          <motion.div
+          <MotionDiv
             initial={{ opacity: 0, y: 24 }}
             whileInView={{ opacity: 1, y: 0 }}
             transition={{ duration: 0.6 }}
@@ -623,7 +621,7 @@ export default function AkomapaNetworkPage() {
                 </p>
               </div>
             </div>
-          </motion.div>
+          </MotionDiv>
         </div>
       </section>
     </>
