+++ src/app/(main)/programs/akomapa-network/page.tsx
@@ -1,11 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
-export const metadata: Metadata = buildPageMetadata(
-  "/programs/akomapa-network",
-);
+export const metadata: Metadata = buildPageMetadata("/programs/akomapa-network");
 
 export default function AkomapaNetworkPage() {
-  return <ClientPage />;
+  return <Content />;
 }
