+++ src/app/(main)/resources/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import ResourcesExplorer from "@/components/resources/ResourcesExplorer";
 
 export const metadata: Metadata = buildPageMetadata("/resources");
 
 export default function ResourcesPage() {
-  return <ClientPage />;
+  return <ResourcesExplorer />;
 }
