+++ src/app/(main)/roadmap/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/roadmap");
 
 export default function RoadmapPage() {
-  return <ClientPage />;
+  return <Content />;
 }
