+++ src/app/(main)/terms/page.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from "next";
 import { buildPageMetadata } from "@/lib/seo";
-import ClientPage from "./ClientPage";
+import Content from "./Content";
 
 export const metadata: Metadata = buildPageMetadata("/terms");
 
 export default function TermsPage() {
-  return <ClientPage />;
+  return <Content />;
 }
