+++ src/components/roadmap/phases.ts
@@ -1,0 +1,55 @@
+export const phases = [
+  {
+    id: 1,
+    title: "Phase 1: Launch & Learn",
+    period: "October 2025–October 2026",
+    focus: "Deliver excellent care, build trust, and lay the legal and strategic foundation.",
+    goal: "Deliver care to our first 500 patients and build strong organizational roots.",
+    color: "bg-[#0097b2]",
+    icon: "Target" as const,
+    achievements: [
+      "Launch free clinic services (screenings, counseling, follow-up)",
+      "Train student volunteers from multiple disciplines",
+      "Establish referral and patient advocacy pathways",
+      "Support NHIS registration for patients",
+      "Begin a home health and follow-up program",
+      "Launch research and data collection team",
+      "Pilot Global Health Leaders Training Program (Yale, UCLA, UCC)",
+      "Launch Akomapa Partner Program (monthly giving community)",
+      "Formalize nonprofit status in Ghana and the U.S.",
+      "Develop early partnerships with local clinics, GHS, and academic institutions",
+    ],
+  },
+  {
+    id: 2,
+    title: "Phase 2: Pilot Sustainability",
+    period: "October 2026–October 2027",
+    focus: "Test social enterprise models and strengthen clinical impact.",
+    goal: "Test bold, community-driven solutions that support long-term care.",
+    color: "bg-[#eeba2b]",
+    icon: "TrendingUp" as const,
+    achievements: [
+      "Launch pilot Akomapa Pharmacy (affordable meds with reinvestment)",
+      "Begin Akomapa Farms test plot (nutrition + income generation)",
+      "Expand NHIS registration and follow-up capacity",
+      "Grow the Global Health Training Program",
+      "Share early research insights to improve clinic practices",
+    ],
+  },
+  {
+    id: 3,
+    title: "Phase 3: Scale & Sustain",
+    period: "October 2027–October 2028",
+    focus: "Build lasting systems, expand impact, and share lessons learned.",
+    goal: "Establish Akomapa as a replicable, sustainable care model in Africa.",
+    color: "bg-[#2F3332]",
+    icon: "Globe" as const,
+    achievements: [
+      "Open a fully functioning Akomapa Pharmacy",
+      "Launch Akomapa Farms & Food Store (revenue supports the clinic)",
+      "Secure long-term institutional sponsors and grant funders",
+      "Present and publish research and program outcomes",
+      "Begin planning for clinic replication across Ghana or the Global South",
+    ],
+  },
+];
